cli/requirements.txt
psycopg2-binary==2.9.10
rich==13.9.4